requirements.txt
Flask==3.0.2
python-pptx==0.6.23
python-docx==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.1 